Feature/add fxapi

+++ src/User.h
@@ -17,31 +17,31 @@ using namespace std;
 
 class User {
 	string name;
-	int balance;
-	int total_owed;
-	int total_owes;
+	double balance;
+	double total_owed;
+	double total_owes;
 	unordered_map<string,vector<string>> transactions;
 	unordered_map<string,int> tagCount;
 
 public:
 	User(string name);
-	User(string name, int balance, int total_owed, int total_owes, unordered_map<string,int> tagCount);
+	User(string name, double balance, double total_owed, double total_owes, unordered_map<string,int> tagCount);
 	void getBalanceDetails();
-	int getBalance();
-	int getTotalOwed();
-	int getTotalOwes();
+	double getBalance();
+	double getTotalOwed();
+	double getTotalOwes();
 	unordered_map<string,int> getTagMap();
 	void balanceOut();
 	int getTotalUnits(vector<string> ratios);
 	//Expense when some other spends and this user is involved
-	void addOwesExpense(string name, int multiplier, int perUnitAmount, string tag);
+	void addOwesExpense(string name, int multiplier, double perUnitAmount, string tag);
 	//Expense when this user spends
-	void addOwedExpense(vector<string> usersToBeSplitWith,vector<string> ratios,int amount,string tag);
+	void addOwedExpense(vector<string> usersToBeSplitWith,vector<string> ratios,double amount,string tag);
 	int simplifyTransactions(string name);
 	void simplify();
 	void expenditureByTag(string tag);
-	void makePaymentToUser(string name, int amount);
-	void receivePaymentFromUser(string name, int amount);
+	void makePaymentToUser(string name, double amount);
+	void receivePaymentFromUser(string name, double amount);
 	virtual ~User(){
 	}
 };
